fix send signature

+++ libs/native/src/main21/java/org/elasticsearch/nativeaccess/jdk/JdkPosixCLibrary.java
@@ -273,7 +273,7 @@ class JdkPosixCLibrary implements PosixCLibrary {
         var segment = nativeBuffer.segment;
         try {
             logger.info("Sending {} bytes to socket", buffer.buffer().remaining());
-            return (long) send$mh.invokeExact(errnoState, sockfd, segment, buffer.buffer().remaining(), flags);
+            return (long) send$mh.invokeExact(errnoState, sockfd, segment, (long) buffer.buffer().remaining(), flags);
         } catch (Throwable t) {
             throw new AssertionError(t);
         }
